Revert

+++ src/util/regexp.h
@@ -167,7 +167,7 @@ class CVC4_PUBLIC String {
    * Returns the maximum length of string representable by this class.
    * Corresponds to the maximum size of d_str.
    */
-  static size_t maxSize();
+  static long maxSize();
  private:
   // guarded
   static unsigned char hexToDec(unsigned char c);
